refactors to make things work properly
